+++ src/Runtime/mmdModel.ts
@@ -53,6 +53,13 @@ export class MmdModel implements IMmdModel {
     public readonly skeleton: IMmdLinkedBoneContainer;
 
     /**
+     * Uint8Array that stores the state of IK solvers
+     *
+     * if `ikSolverState[MmdModel.sortedRuntimeBones[i].ikSolverIndex]` is 0, IK solver of `MmdModel.sortedRuntimeBones[i]` is disabled and vice versa
+     */
+    public readonly ikSolverStates: Uint8Array;
+
+    /**
      * The morph controller of this model
      *
      * The `MmdMorphController` not only wrapper of `MorphTargetManager` but also controls the CPU bound morphs (bone, material, group)
@@ -125,6 +132,15 @@ export class MmdModel implements IMmdModel {
         };
         this.mesh = runtimeMesh;
         this.skeleton = skeleton;
+
+        {
+            let ikSolverCount = 0;
+            const bonesMetadata = mmdMetadata.bones;
+            for (let i = 0; i < bonesMetadata.length; ++i) {
+                if (bonesMetadata[i].ik !== undefined) ikSolverCount += 1;
+            }
+            this.ikSolverStates = new Uint8Array(ikSolverCount).fill(1);
+        }
 
         // If you are not using MMD Runtime, you need to update the world matrix once. it could be waste of performance
         skeleton.prepare();
@@ -287,15 +303,7 @@ export class MmdModel implements IMmdModel {
      */
     public resetState(): void {
         this.morph.resetMorphWeights();
-
-        const sortedBones = this._sortedRuntimeBones;
-        for (let i = 0; i < sortedBones.length; ++i) {
-            const bone = sortedBones[i];
-
-            if (bone.ikSolver !== null) {
-                bone.ikSolver.enabled = true;
-            }
-        }
+        this.ikSolverStates.fill(1);
     }
 
     /**
@@ -366,11 +374,9 @@ export class MmdModel implements IMmdModel {
                 bone.updateWorldMatrix();
             }
 
-            if (bone.ikSolver !== null) {
+            if (bone.ikSolver !== null && this.ikSolverStates[i] !== 0) {
                 bone.ikSolver.solve();
-                if (bone.ikSolver.enabled) {
-                    bone.updateWorldMatrix();
-                }
+                bone.updateWorldMatrix();
             }
         }
 
@@ -392,6 +398,7 @@ export class MmdModel implements IMmdModel {
             runtimeBones.push(new MmdRuntimeBone(bones[i], boneMetadata));
         }
 
+        let ikSolverCount = 0;
         for (let i = 0; i < bonesMetadata.length; ++i) {
             const boneMetadata = bonesMetadata[i];
             const bone = runtimeBones[i];
@@ -421,9 +428,11 @@ export class MmdModel implements IMmdModel {
                 const targetBoneIndex = ikMetadata.target;
                 if (0 <= targetBoneIndex && targetBoneIndex < runtimeBones.length) {
                     const ikSolver = bone.ikSolver = new IkSolver(
+                        ikSolverCount,
                         bone,
                         runtimeBones[targetBoneIndex]
                     );
+                    ikSolverCount += 1;
                     ikSolver.iteration = ikMetadata.iteration;
                     ikSolver.limitAngle = ikMetadata.rotationConstraint;
                     for (let j = 0; j < ikMetadata.links.length; ++j) {